+++ bin/data/toml_config.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:path/path.dart' show canonicalize;
 import 'package:toml/toml.dart';
 
 import '../extensions/chalk_ext.dart';
@@ -21,11 +22,11 @@ class TomlConfig {
     final document = TomlDocument.fromMap({
       'hooks': {
         'successful_check': [
-          {'create_file': TomlBasicString(config.tcFile.path)},
+          {'create_file': TomlBasicString('{workspace_dir}\\.reds-ready')},
         ],
       },
     });
-    final file = config.rlsFile;
+    final file = config.defaultRLSFile;
 
     try {
       file.writeAsStringSync(document.toString());
@@ -39,8 +40,10 @@ class TomlConfig {
   factory TomlConfig.fromDocument(TomlDocument document) {
     try {
       final toml = document.toMap();
-      final path = toml['hooks']?['successful_check']?[0]?['create_file'] ?? '';
+      String path = toml['hooks']?['successful_check']?[0]?['create_file'] ?? '';
 
+      path = path.replaceFirst('{workspace_dir}', Directory.current.path);
+      path = canonicalize(path);
       return TomlConfig(path: path);
     } catch (error) {
       return TomlConfig(path: '');